Fix variable types according to assembly use
* linux/dev/arch/i386/kernel/irq.c (intr_count): Set to int type.
* linux/dev/kernel/softirq.c (bh_active, bh_mask): Likewise.
(linux_soft_intr: active, mask, left): Likewise
* linux/src/include/linux/interrupt.h (bh_active, bh_mask): Likewise.
* linux/src/kernel/softirq.c (intr_count, bh_active, bh_mask): Likewise.
(do_bottom_half: active, mask, left): Likewise

+++ linux/dev/arch/i386/kernel/irq.c
@@ -72,7 +72,7 @@ spl_t linux_intr_pri;
 /*
  * Flag indicating an interrupt is being handled.
  */
-unsigned long intr_count = 0;
+unsigned int intr_count = 0;
 
 /*
  * List of Linux interrupt handlers.
