+++ linux/src/include/linux/interrupt.h
@@ -17,8 +17,8 @@ struct irqaction {
 extern unsigned long intr_count;
 
 extern int bh_mask_count[32];
-extern unsigned long bh_active;
-extern unsigned long bh_mask;
+extern unsigned int bh_active;
+extern unsigned int bh_mask;
 extern void (*bh_base[32])(void);
 
 asmlinkage void do_bottom_half(void);
